Adding MultiColorTriangle Classes
Adding MultiColorTriangle Classes. The classes under Apps\Sandcastle
directory are examples of how to use the MultiColorTriangle Classes and
everything under Source\Core are supporting classes for the API. The
classes under Source\Scene and Source\Shaders define the WebGL code.
Finally, The code under Source\Workers is worker call that is called via
a thread to asynchronously render each individual polygon. Some testing
needs to be done to verify that the worker method is functioning
properly. However, it's not a requirement for initial functionality.